Fix mitm_slave to return struct slave *
@p-jean says:
====================
The macro mitm_slave() is used as if it returns a point to 'struct slave'. It returns a pointer to 'struct net_device' instead.

The type was changed on Jan 13 with a commit called 'Use netdevice where appropriate'.

In mitm_master_upper_dev_link(), slave_dev rather than slave is now used as the 3rd argument of netdev_master_upper_dev_link().

A point to struct slave is a pointer to a pointer to the slave_dev, so not the same as a point to net_device.
====================

Fixes #3.